requirements.txt
Django==5.0.1
django-rest-framework==0.1.0
python-dotenv==1.0.0